cleaner decode code
